+++ app/src/main/java/org/fullmetalfalcons/androidscouting/activities/AboutActivity.java
@@ -1,8 +1,10 @@
-package org.fullmetalfalcons.androidscouting;
+package org.fullmetalfalcons.androidscouting.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+
+import org.fullmetalfalcons.androidscouting.R;
 
 public class AboutActivity extends AppCompatActivity {
