+++ app/src/main/res/layout/activity_retrieve_data.xml
@@ -59,7 +59,9 @@
                 android:inputType="number"
                 android:ems="10"
                 android:id="@+id/retrieve_team_num"
-                android:maxLength="4"/>
+                android:maxLength="4"
+                android:imeOptions="actionDone"
+                android:singleLine="true"/>
 
             <Button
                 android:layout_width="wrap_content"
@@ -122,7 +124,9 @@
             android:id="@+id/editText2"
             android:layout_gravity="center_horizontal"
             android:hint="Value"
-            android:textAlignment="center" />
+            android:textAlignment="center"
+            android:imeOptions="actionDone"
+            android:singleLine="true"/>
 
         <Button
             android:layout_width="wrap_content"
